ttas/requirements.txt
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.1.0
giotto-tda>=0.6.0
ripser>=0.6.0
persim>=0.3.1
kmapper>=2.0.1
plotly>=5.10.0
dash>=2.6.0
umap-learn>=0.5.3
pydiffmap>=0.2.0
fredapi>=0.5.0
osmnx>=1.2.2
econml>=0.14.0
dowhy>=0.8
# tigramite # can be tricky to install, uncomment if needed
networkx
matplotlib
statsmodels
